requirements.txt
selenium==4.21.0
customtkinter==5.2.1
pandas==2.2.2
python-dotenv==1.0.1
 